Basic implementation of multithreaded downloading. Need to change function calls into a full exec.

+++ queue.h
@@ -2,7 +2,7 @@
 #define QUEUE_H_INCLUDED
 
 struct QNode{ 
-    int key;
+    long key;
     struct QNode *next;
 };
 struct Queue{
